src/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
notion-client==2.2.1
openai==1.14.3
outcome==1.3.0.post0
pydantic==2.6.4
pydantic_core==2.16.3
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
selenium==4.18.1
setuptools==69.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
undetected-chromedriver==3.5.5
urllib3==2.2.1
websockets==12.0
wsproto==1.2.0
